backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# RAG packages
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
tiktoken==0.5.2
nltk==3.8.1
numpy==1.26.3

# LLM providers (install as needed)
langchain-google-genai>=0.0.5
openai==1.10.0
# groq==0.4.1
# langchain-groq
# ollama
